KAFKA-16223 Migrate KafkaConfigBackingStoreTest to Mockito (#16126)
Reviewers: Chia-Ping Tsai <chia7712@gmail.com>
